:notebook: extract save participant into contrib, prepare for more particpants. (#176683)
